ANDROID: cpu_cooling: Remove the dependency on the freq_table if not used
Each CPU cooling device is associated with a custom frequency table
which holds both power and frequency values. When using the
cpufreq_cooling_ops, the power values are not used.

Remove the dependency on the ad-hoc frequency table when the
cpufreq_frequency table of the cpufreq_policy can be used instead.
This should help transitioning IPA to using the newly introduced Energy
Model framework without risking to impact those who don't use the model.

Change-Id: I4c3f4a0aff6fc158bc4f265f01d367c8d816524f
Signed-off-by: Quentin Perret <quentin.perret@arm.com>