sched/cpufreq: Make utilization aggregation parametric
Schedutil offers an API to aggregate PELT signals of CFS, RT, DL and
IRQ. However, the way signals are aggregated depends on the final usage
of the aggregated value. For frequency selection, filtering policies
could be applied; for example we go to max frequency when RT tasks are
active. The Energy Aware Scheduler (EAS) instead needs to know the
expected CPU busy time to assess the energy impact of scheduling
decisions. Right now this dual behavior is obtained by passing in a
"request type", which however could make the code more complex in case
additional policies needs to be added.

Let's re-organize the code of the utilization aggregation function to use
a more generic pattern. Do that by moving the core logic of:
   sched/cpufreq_schedutil::schedutil_freq_util()
into:
   sched/sched.h::__cpu_util()
and refactor it to exploit constants propagation at compile time to
provide the specialized code version required by the different usage
scenarios.

The new __cpu_util() has a set of parameters which allow to specify the
utilization contributions to use.  When all the parameters are zero, it
computes a utilization suitable for frequency selection using the actual
PELT contributions and applying the required filtering policies on top.

When instead some utilization contributions are non zero, filtering
policies are never applied then the specified contributions are used
instead of the corresponding class utilization, for the computation of
the (expected) aggregated utilization.

This second mode allows to implement filtering policies at the calling
side thus supporting a cleaner implementation of scheduling scenarios
evaluation and comparison.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>