sched/cpufreq: Make utilization aggregation parametric
Schedutil aggregates the PELT signals of CFS, RT, DL and IRQ in order
to decide which frequency to request. Energy Aware Scheduling (EAS)
needs to be able to predict those requests to assess the energy impact
of scheduling decisions. However, the PELT signals aggregation is only
done in schedutil for now, hence making it hard to synchronize it with
EAS, and signals aggregation required could be different for frequency
selection or tasks placement decisions.

To address these issues, let's move the core logic of sugov_get_util()
into kernel/sched/sched.h::__cpu_util() and refactor it to provide
different code version by constants propagation at compile time, thus
meeting by "specialization" the usage requirements of different
scheduler components. For example, frequency selection and energy
estimation needs to deal with RT and DL signals slightly differently.

The new __cpu_util() has a set of parameters which allow to specify the
utilization contributions to use.  When all the parameters are zero, it
computes a utilization suitable for frequency selection. When some
utilization contributions are non zero, then the specified contribution
is used instead of the corresponding class utilization, for the
computation of the (expected) aggregated utilization.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>