sched/core: uclamp: Add system default clamps
Tasks without a user-defined clamp value are considered not clamped
and by default their utilization can have any value in the range
[0..SCHED_CAPACITY_SCALE].

Tasks with a user-defined clamp value are allowed to request any value
in that range, and we unconditionally enforce the required clamps.
However, a "System Management Software" could be interested in limiting
the range of clamp values allowed for all tasks.

Add a privileged interface to define a system default configuration via:

  /proc/sys/kernel/sched_uclamp_util_{min,max}

which works as an unconditional clamp range restriction for all tasks.

All task specific values are forced to be not greater then the
corresponding system default clamp value.

Signed-off-by: Patrick Bellasi <patrick.bellasi@arm.com>
Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>

---
Changes in v7:
 Message-ID: <20190124123009.2yulcf25ld66popd@e110439-lin>
 - set sched_uclamp_util_min=1024 by default and implement a "nice"
   policy for system default: task-specific util_min can only be
   not greater then the configured system default value.
 - update uclamp_effective_get() to enfore a "nice" policy
 - add a comment to document the "lazy" update in sched_uclamp_handler()