HACK: module: Middle-finger to the version check
There are issues on XZ2 because the vendor partition expects a specific
header version, and we want to change those headers, but we can't update
that fscking partition.

To 'workaround' the issue, just remove the version check. Yes, this is
atrocious, and that will result without a doubt in borked devices, but
there isn't much we can do about it.

The consequences are, all new members to structs defined in include/
must be added _at the end_ of the structs, and the offsets _must_
be kept the same.

This honnestly just plain sucks

Not-signed-off-by: Quentin Perret <quentin.perret@arm.com>