WiP: sched: Add some doc for EAS
Because Energy Aware Scheduling (EAS) relies on several subsystems,
overloads multiple mechanisms of the scheduler, and assumes a number of
things about the platform, it isn't that intuitive to use, and even
less intuitive to fix.

So, let's try to explain in a doc file how all of that works, and what
are the bits and pieces required to get it up and running.

Signed-off-by: Quentin Perret <quentin.perret@arm.com>