ANDROID: thermal: cpu_cooling: Remove unecessary dependency on freq_table
When CPU cooling devices are registered without a power model, their
states are still stored in a struct freq_table, which includes null
power values. However, the CPU cooling framework already has a
dependency on the frequency table managed by CPUFreq, which basically
holds a copy of the data.

In order to avoid the duplication, make the cpufreq_cooling_ops use
the CPUFreq frequency table instead of the internal copy, hence reducing
the the amount of duplicated data.

Change-Id: If4eabd6ffa6e257db812ea70860276949029d175
Signed-off-by: Quentin Perret <quentin.perret@arm.com>