ANDROID: sched: Replace the EAS static key by a sched_feat
Now that the Energy Model framework is used not only by the scheduler
but also by the thermal subsystem, there is no single tunable left to
disable EAS by hand. As a result, it is not possible to use IPA on a
big.LITTLE platform without enabling EAS automatically.

Since some users are interested in using IPA for performance reasons
and not so concerned about energy consumption, replace the automatically
enabled sched_energy_present static key by a sched_feat.

Since all of the EAS code paths are guarded by a double check on the
static key (now sched_feat) and the presence of an EM for the current
root domain, there is no risk of accidentally entering them by mistake.
It is still safe to enable the static key even without an EM -- that
will result in loosing a very minor optimization in select_task_rq_fair()
and in the load balance path.

The new sched feature is called ENERGY_AWARE in order to be aligned with
the historical EAS implementation.

Change-Id: I6f6d17f5beccd054b7b89478a1e792e37d14245e
Signed-off-by: Quentin Perret <quentin.perret@arm.com>